backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai
python-dotenv
networkx
sqlalchemy
bcrypt
python-jose
